+++ lib/spid/rack/sso.rb
@@ -57,8 +57,26 @@ module Spid
           end
         end
 
-        def valid_request?
+        def valid_get?
+          request.get? &&
+            Spid.configuration.acs_binding == Spid::BINDINGS_HTTP_REDIRECT
+        end
+
+        def valid_post?
+          request.post? &&
+            Spid.configuration.acs_binding == Spid::BINDINGS_HTTP_POST
+        end
+
+        def valid_http_verb?
+          valid_get? || valid_post?
+        end
+
+        def valid_path?
           request.path == Spid.configuration.acs_path
+        end
+
+        def valid_request?
+          valid_path? && valid_http_verb?
         end
 
         def sso_response